+++ femmt/Functions.py
@@ -13,6 +13,7 @@ import time
 import warnings
 from typing import Union, List, Tuple, Dict
 from matplotlib import pyplot as plt
+from femmt.Enumerations import ConductorType
 
 # Third parry libraries
 import gmsh
@@ -260,6 +261,50 @@ def wire_material_database() -> Dict:
 
     return wire_material
 
+def cost_material_database() -> Dict:
+    """
+    Returns costs for core and winding.
+    This is splitted in material and fabrication costs.
+    Both, material and fabrication costs have a euro_per_kilogram and a euro_per_unit (fixcosts) price.
+
+    Source: R. Burkart and J. Kolar 'Component Cost Models for Multi-Objective Optimizations of Switched-Mode Power Converter'
+    2013.
+
+    These are outdated prices (year 2013). Update needed in future.
+    """
+
+    cost_database = {}
+    cost_database["core_euro_per_kilogram"] = {"ferrite": 5.5,
+                                               "amorphous": 16,
+                                               "nanocristalline": 23,
+                                               "high_si_steel": 12,
+                                               "goes": 2.5}
+    cost_database["winding_material_euro_per_kilogram"] = {ConductorType.RoundSolid.name: 10,
+                                                  "flat": 10,
+                                                  ConductorType.RectangularSolid.name: 20,
+                                                  ConductorType.RoundLitz.name: -1}
+
+    cost_database["winding_material_euro_per_unit"] = {ConductorType.RoundSolid.name: 1,
+                                                  "flat": 2,
+                                                  ConductorType.RectangularSolid.name: 2,
+                                                  ConductorType.RoundLitz.name: 1}
+
+    cost_database["winding_fabrication_euro_per_kilogram"] = {ConductorType.RoundSolid.name: 7,
+                                                  "flat": 21,
+                                                  ConductorType.RectangularSolid.name: 14,
+                                                  ConductorType.RoundLitz.name: 7}
+    cost_database["winding_fabrication_euro_per_unit"] = {ConductorType.RoundSolid.name: 2,
+                                                  "flat": 4,
+                                                  ConductorType.RectangularSolid.name: 2.5,
+                                                  ConductorType.RoundLitz.name: 2}
+
+    cost_database["winding_material_euro_per_kilogram_for_litz"] = {"sigma_numerator": 15,
+                                                                    "sigma_denumerator": 0.45}
+
+    cost_database["gross_margin"] = 0.25
+
+
+    return cost_database
 
 def pm_core_inner_diameter_calculator(inner_core_diameter: float, hole_diameter: float) -> float:
     """
@@ -1067,6 +1112,90 @@ def set_silent_status(s: bool):
 def femmt_print(text):
     if not silent:
         print(text)
+
+def cost_function_core(core_weight: float, core_type: str = "ferrite") -> float:
+    """
+    Calculates core material costs depending on material and weight.
+
+    :param core_weight: core weight in kg
+    :type core_weight: float
+    :param core_type: core type. Can be "ferrite", "amorphous", "nanocristalline", "high_si_steel", "goes"
+    :type core_type: str
+    :return: costs of core in euro
+    :rtype: float
+    """
+    cost_database = cost_material_database()
+    sigma_core = cost_database["core_euro_per_kilogram"][core_type]
+
+    return sigma_core * core_weight
+
+
+def cost_function_winding(winding_weight_list: List[float], winding_type: List[str], single_strand_cross_section_list: List[float] = []):
+    """
+    Calculates single winding material and fabrication costs depending on winding-type and weight
+
+    :param winding_weight_list: winding weight in kg in list-form
+    :type winding_weight_list: List[float]
+    :param winding_type: winding type. Must fit to enum-names in ConductorType-Enum
+    :type winding_type: List[str]
+    :param single_strand_cross_section_list: single strand cross section in list-form
+    :type single_strand_cross_section_list: List[float]
+    :return: winding cost of single winding
+    :rtype: float
+    """
+    cost_database = cost_material_database()
+    winding_cost_list = []
+
+    for winding_count, winding_weight in enumerate(winding_weight_list):
+        # material cost (per kilogram and per unit)
+        sigma_material_winding_euro_per_kilogram = cost_database["winding_material_euro_per_kilogram"][winding_type[winding_count]]
+        if sigma_material_winding_euro_per_kilogram == -1:
+            # case for special litz wire calculation. Additional data is loaded from cost_database.
+            sigma_material_winding_euro_per_kilogram = cost_database["winding_material_euro_per_kilogram_for_litz"]["sigma_numerator"] / (single_strand_cross_section_list[winding_count] * 1e6 + cost_database["winding_material_euro_per_kilogram_for_litz"]["sigma_denumerator"])
+
+        winding_material_euro_per_unit = cost_database["winding_material_euro_per_unit"][winding_type[winding_count]]
+
+        winding_material_cost = sigma_material_winding_euro_per_kilogram * winding_weight + winding_material_euro_per_unit
+
+        # fabrication cost (per kilogram and per unit)
+        sigma_fabrication_euro_per_kilogram = cost_database["winding_fabrication_euro_per_kilogram"][winding_type[winding_count]]
+        fabrication_material_euro_per_unit = cost_database["winding_fabrication_euro_per_unit"][winding_type[winding_count]]
+
+        winding_fabrication_cost = sigma_fabrication_euro_per_kilogram * winding_weight + fabrication_material_euro_per_unit
+
+        winding_cost_list.append(winding_material_cost + winding_fabrication_cost)
+
+    return winding_cost_list
+
+
+
+def cost_function_total(core_weight: float, core_type: str, winding_weight_list: List[float], winding_type_list: List[str]) -> float:
+    """
+    Calculates the total costs for a inductive element.
+    This includes material costs for core and winding, fabrication costs for core and winding and manufacturer margin
+
+    :param core_weight: core weight in kg
+    :type core_weight: float
+    :param core_type: core type. Can be "ferrite", "amorphous", "nanocristalline", "high_si_steel", "goes"
+    :type core_type: str
+    :param winding_weight_list: winding weight in kg
+    :type winding_weight_list: float
+    :param winding_type_list: winding type in list-form. Must fit to enum-names in ConductorType-Enum
+    :type winding_type_list: List[str]
+    :return: total costs for inductive element
+    :rtype: float
+    """
+    cost_database = cost_material_database()
+
+    cost_core = cost_function_core(core_weight, core_type)
+
+
+    cost_winding_list = cost_function_winding(winding_weight_list, winding_type_list)
+    cost_winding = sum(cost_winding_list)
+
+    total_cost_including_margin = 1 / (1 - cost_database["gross_margin"]) * (cost_core + cost_winding)
+
+    return total_cost_including_margin
 
 if __name__ == '__main__':
     # TODO Relative path